moved image controllers to the sites blueprint
