+++ templates/items/view_item.html
@@ -21,8 +21,8 @@
     <div class="row images">
         {% for image in item.images %}
             <div class="col-sm-12 col-md-4">
-                <a href="{{ url_for('items.get_image', filename=image) }}">
-                    <img class="thumbnail" src="{{ url_for('items.get_thumbnail', filename=image) }}">
+                <a href="{{ url_for('main.get_image', filename=image) }}">
+                    <img class="thumbnail" src="{{ url_for('main.get_thumbnail', filename=image) }}">
                 </a>
             </div>
         {% endfor %}
